In this episode: Zero forgets an f

+++ src/seascape.c
@@ -112,7 +112,7 @@ inline f32 hash(guVec2 p) {
 	guVec2 hv = (guVec2) { 127.1f, 311.7f };
 	float h = guVec2Dot(p, hv);
 	float n = sinf(h) * 43758.5453123;
-	return n - floor(n);
+	return n - floorf(n);
 }
 
 inline f32 noise(guVec2 p) {
